Move i686 memory types and functions to C headers This allows the stivale2 bootloader stage to access the structs without C++ inclusion jank. That way additional classes can be added to `Memory.i686.hpp` without the `#ifdefs` becoming a nightmare to deal with. Plus, traditionally, C++ includes C headers, not the other way around.
